refactor(api): migrate profile collection campaign result to its bounded context
