+++ api/src/prescription/campaign/application/campaign-results-route.js
@@ -1,5 +1,5 @@
 import Joi from 'joi';
-import { campaignParticipationController } from './campaign-participation-controller.js';
+import { campaignResultsController } from './campaign-results-controller.js';
 import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
 import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
 
@@ -24,12 +24,38 @@ const register = async function (server) {
             'page[size]': Joi.number().integer().empty(''),
           }),
         },
-        handler: campaignParticipationController.findAssessmentParticipationResults,
+        handler: campaignResultsController.findAssessmentParticipationResults,
         notes: [
           '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
             "- Récupération des résultats d'une campagne d'évaluation",
         ],
-        tags: ['api', 'campaign-assessment-participation-result-minimal'],
+        tags: ['api', 'campaign-results-assessment'],
+      },
+    },
+    {
+      method: 'GET',
+      path: '/api/campaigns/{id}/profiles-collection-participations',
+      config: {
+        pre: [{ method: securityPreHandlers.checkAuthorizationToAccessCampaign }],
+        validate: {
+          params: Joi.object({
+            id: identifiersType.campaignId,
+          }),
+          query: Joi.object({
+            'filter[divisions][]': [Joi.string(), Joi.array().items(Joi.string())],
+            'filter[groups][]': [Joi.string(), Joi.array().items(Joi.string())],
+            'filter[search]': Joi.string().empty(''),
+            'filter[certificability]': Joi.string().empty(''),
+            'page[number]': Joi.number().integer().empty(''),
+            'page[size]': Joi.number().integer().empty(''),
+          }),
+        },
+        handler: campaignResultsController.findProfilesCollectionParticipations,
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+            "- Récupération des profils collectés d'une campagne par son id",
+        ],
+        tags: ['api', 'campaign-results-profiles-collection'],
       },
     },
   ]);
