+++ api/src/prescription/campaign/domain/read-models/CampaignProfilesCollectionParticipationSummary.js
