+++ api/src/prescription/campaign/domain/usecases/find-campaign-profiles-collection-participation-summaries.js
