+++ api/src/prescription/campaign/routes.js
@@ -1,7 +1,7 @@
 import * as campaignAdministration from './application/campaign-administration-route.js';
 import * as campaignDetail from './application/campaign-detail-route.js';
-import * as campaignParticipation from './application/campaign-participation-route.js';
+import * as campaignResults from './application/campaign-results-route.js';
 
-const campaignRoutes = [campaignAdministration, campaignDetail, campaignParticipation];
+const campaignRoutes = [campaignAdministration, campaignDetail, campaignResults];
 
 export { campaignRoutes };
